+++ lib/src/dictionary/_index.dart
@@ -4,6 +4,6 @@
 
 export 'dictionary.dart';
 export 'term_properties.dart';
-export 'part_of_speech.dart';
-export 'term_variant.dart';
+export '../part_of_speech.dart';
+export 'term_definition.dart';
 // export '../../../test/oxford_dictionary.dart';